+++ src/frontend/src/components/buttons/ScanButton.tsx
