+++ src/frontend/src/components/forms/InstanceOptions.tsx
@@ -6,7 +6,7 @@ import { IconCheck } from '@tabler/icons-react';
 import { useServerApiState } from '../../states/ApiState';
 import { useLocalState } from '../../states/LocalState';
 import { HostList } from '../../states/states';
-import { EditButton } from '../items/EditButton';
+import { EditButton } from '../buttons/EditButton';
 import { HostOptionsForm } from './HostOptionsForm';
 
 export function InstanceOptions({
